Run migrations on first creation, fix a bunch of panics

+++ src/input.rs
@@ -78,16 +78,20 @@ pub fn handle(state: &mut State<'_>) -> Result<(), std::io::Error> {
                 KeyCode::Char('g') => column.select_first_task(),
                 KeyCode::Char('G') => column.select_last_task(),
                 KeyCode::Char('H') => {
-                    project.move_task_previous_column();
-                    let col = project.get_selected_column();
-                    let t = col.get_selected_task().unwrap();
-                    db::move_task_to_column(&state.db_conn, &t, &col);
+                    if !column.tasks.is_empty() {
+                        project.move_task_previous_column();
+                        let col = project.get_selected_column();
+                        let t = col.get_selected_task().unwrap();
+                        db::move_task_to_column(&state.db_conn, &t, &col);
+                    }
                 }
                 KeyCode::Char('L') => {
-                    project.move_task_next_column();
-                    let col = project.get_selected_column();
-                    let t = col.get_selected_task().unwrap();
-                    db::move_task_to_column(&state.db_conn, &t, &col);
+                    if !column.tasks.is_empty() {
+                        project.move_task_next_column();
+                        let col = project.get_selected_column();
+                        let t = col.get_selected_task().unwrap();
+                        db::move_task_to_column(&state.db_conn, &t, &col);
+                    }
                 }
                 KeyCode::Char('J') => {
                     if column.move_task_down() {
@@ -108,8 +112,10 @@ pub fn handle(state: &mut State<'_>) -> Result<(), std::io::Error> {
                     state.task_edit_state = column.get_task_state_from_curr_selected_task()
                 }
                 KeyCode::Char('D') => {
-                    db::delete_task(&state.db_conn, column.get_selected_task().unwrap());
-                    column.remove_task();
+                    if !column.tasks.is_empty() {
+                        db::delete_task(&state.db_conn, column.get_selected_task().unwrap());
+                        column.remove_task();
+                    }
                 }
                 _ => {}
             },
